# docs/api/edgechat-v1.openapi.yaml
openapi: 3.1.0
info:
  title: EdgeChat Mobile API
  version: 1.0.0
  description: Stable HTTP contract shared by native clients and EdgeChat servers.
servers:
  - url: https://chat.example.com/api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      security: []
      operationId: loginDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Device session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileAuth"
        "401":
          $ref: "#/components/responses/ApiError"
  /auth/refresh:
    post:
      security: []
      operationId: refreshDeviceSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken, installationId]
              properties:
                refreshToken: { type: string }
                installationId: { type: string, format: uuid }
      responses:
        "200":
          description: Rotated device tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileAuth"
        "401":
          $ref: "#/components/responses/ApiError"
  /auth/logout:
    post:
      operationId: logoutDevice
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/ApiError" }
...
  /rooms/{kind}/{roomId}/messages:
    get:
      operationId: listRoomMessages
      parameters:
        - $ref: "#/components/parameters/RoomKind"
        - $ref: "#/components/parameters/RoomId"
        - name: before
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: Canonical message page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomMessages"
    post:
      operationId: sendRoomMessage
      parameters:
        - $ref: "#/components/parameters/RoomKind"
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientMessageId, content]
              properties:
                clientMessageId: { type: string, format: uuid }
                content: { type: string }
                attachment: { oneOf: [{ $ref: "#/components/schemas/Attachment" }, { type: "null" }] }
                mentionUserIds:
                  type: array
                  items: { type: integer }
                replyMessageId: { type: [integer, "null"], minimum: 1 }
      responses:
        "200": { $ref: "#/components/responses/MessageSubmission" }
        "201": { $ref: "#/components/responses/MessageSubmission" }
  /rooms/{kind}/{roomId}/messages/{messageId}:
    delete:
      operationId: deleteRoomMessage
      parameters:
        - $ref: "#/components/parameters/RoomKind"
        - $ref: "#/components/parameters/RoomId"
        - name: messageId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
  /rooms/{kind}/{roomId}/sync:
    get:
      operationId: syncRoom
      parameters:
        - $ref: "#/components/parameters/RoomKind"
        - $ref: "#/components/parameters/RoomId"
        - name: cursor
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: Ordered message and deletion events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSync"
        "409":
          description: Cursor expired; reload the current room snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /rooms/{kind}/{roomId}/read:
    post:
      operationId: markRoomRead
      parameters:
        - $ref: "#/components/parameters/RoomKind"
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messageId: { type: [integer, "null"] }
      responses:
        "200":
          description: Read cursor
  /uploads:
    post:
      operationId: uploadAttachment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, clientUploadId]
              properties:
                file: { type: string, format: binary }
                clientUploadId: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/Upload" }
        "201": { $ref: "#/components/responses/Upload" }
  /realtime/tickets:
    post:
      operationId: createRealtimeTicket
      description: |
        Issues a single-use ticket for the WebSocket upgrade endpoint.
        The current D1-only deployment has no Durable Objects, so
        `GET /api/v1/realtime/ws` answers `501 realtime_unsupported` and the
        ticket is never consumed. Clients should poll
        `GET /api/v1/rooms/{kind}/{id}/sync` instead and treat
        `features.realtimeTickets: false` as "do not attempt a socket".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope]
              properties:
                scope: { type: string, enum: [inbox, room] }
                roomKind: { type: string, enum: [public, private, dm] }
                roomId: { type: integer }
      responses:
        "200":
          description: Single-use ticket valid for 60 seconds
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expiresAt]
                properties:
                  ticket: { type: string }
                  expiresAt: { type: string, format: date-time }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    RoomKind:
      name: kind
      in: path
      required: true
      schema: { type: string, enum: [public, private, dm] }
    RoomId:
      name: roomId
      in: path
      required: true
      schema: { type: integer }
  responses:
    ApiError:
      description: Stable client-visible error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Ok:
      description: Successful command
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, const: true }
    MessageSubmission:
      description: Canonical message; created is false for an idempotent replay
      content:
        application/json:
          schema:
            type: object
            required: [created, message]
            properties:
              created: { type: boolean }
              message: { $ref: "#/components/schemas/Message" }
    Upload:
      description: Canonical uploaded file; created is false for an idempotent replay
      content:
        application/json:
          schema:
            type: object
            required: [created, file]
            properties:
              created: { type: boolean }
              file: { $ref: "#/components/schemas/Attachment" }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    LoginRequest:
      type: object
      required: [username, password, device]
      properties:
        username: { type: string }
        password: { type: string }
        device:
          type: object
          required: [installationId, name, appVersion]
          properties:
            installationId: { type: string, format: uuid }
            name: { type: string }
            appVersion: { type: string }
    MobileAuth:
      type: object
      required: [accessToken, accessTokenExpiresAt, refreshToken, refreshTokenExpiresAt, session]
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string }
        refreshToken: { type: string }
        refreshTokenExpiresAt: { type: string }
        session: { $ref: "#/components/schemas/Session" }
    Session:
      type: object
      required: [userId, username, displayName, avatarUrl, isAdmin, sessionVersion]
      properties:
        userId: { type: integer }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        bio:
          type: string
          description: Present on Web login, profile update and session restore; omitted by older/device session producers.
        isAdmin: { type: boolean }
        sessionVersion: { type: integer }
        deviceSessionId: { type: string }
...
    Attachment:
      type: object
      required: [key, name, type, size, url]
      properties:
        key: { type: string }
        name: { type: string }
        type: { type: string }
        size: { type: integer }
        url: { type: string }
        kind: { type: string, enum: [voice, audio] }
        durationMs: { type: integer, minimum: 0 }
        waveform:
          type: array
          maxItems: 64
          items: { type: integer, minimum: 0, maximum: 100 }
    Sender:
      type: object
      required: [kind, id, username, displayName, avatarUrl, source]
      properties:
        kind: { type: string, enum: [local, external] }
        id: { oneOf: [{ type: integer }, { type: string }] }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        source: { type: string }
    Mention:
      type: object
      required: [userId, username, displayName]
      properties:
        userId: { type: integer }
        username: { type: string }
        displayName: { type: string }
    ReplyPreview:
      type: object
      required: [id, deleted]
      properties:
        id: { type: integer }
        deleted: { type: boolean }
        content: { type: string }
        sender: { $ref: "#/components/schemas/Sender" }
        attachment: { oneOf: [{ $ref: "#/components/schemas/Attachment" }, { type: "null" }] }
    Message:
      type: object
      required: [id, content, createdAt, source, sender, attachment]
      properties:
        id: { type: integer }
        clientMessageId: { type: string, format: uuid }
        content: { type: string }
        createdAt: { type: string }
        source: { type: string }
        sender: { $ref: "#/components/schemas/Sender" }
        attachment: { oneOf: [{ $ref: "#/components/schemas/Attachment" }, { type: "null" }] }
        mentionUserIds: { type: array, items: { type: integer } }
        mentions: { type: array, items: { $ref: "#/components/schemas/Mention" } }
        replyToMessageId: { type: integer }
        replyTo: { $ref: "#/components/schemas/ReplyPreview" }
...
    RoomMessages:
      type: object
      required: [room, messages, syncCursor]
      properties:
        room:
          type: object
          required: [id, kind, name, description]
          properties:
            id: { type: integer }
            kind: { type: string }
            name: { type: string }
            description: { type: string }
        messages: { type: array, items: { $ref: "#/components/schemas/Message" } }
        syncCursor: { type: integer }
    RoomSync:
      type: object
      required: [events, nextCursor, hasMore]
      properties:
        events:
          type: array
          items:
            type: object
            required: [sequence, type]
            properties:
              sequence: { type: integer }
              type: { type: string, enum: [message, message_deleted] }
              messageId: { type: integer }
              message: { $ref: "#/components/schemas/Message" }
              createdAt: { type: string }
        nextCursor: { type: integer }
        hasMore: { type: boolean }
        # 后加的可选字段：老客户端忽略即可。
        typing:
          type: array
          items:
            type: object
            required: [userId]
            properties:
              userId: { type: integer }
              displayName: { type: string }
              avatarUrl: { type: string }